Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -1,6 +1,6 @@
 # include<stdio.h>
 int main () {
-    int side1, side2, side3;
+    float side1, side2, side3;
     scanf("%f %f %f", &side1, &side2, &side3);
     if((side1 + side2 > side3) && (side2 + side3 > side1) && (side3 + side1 > side2)){
         printf("Valid\n");
